fix: gettext calls
