+++ lib/extension/window.js
@@ -25,6 +25,7 @@ import St from "gi://St";
 
 // Gnome Shell imports
 import { Overview } from "resource:///org/gnome/shell/ui/overview.js";
+import { PACKAGE_VERSION } from "resource:///org/gnome/shell/misc/config.js";
 
 // Shared state
 import { Logger } from "../shared/logger.js";
@@ -41,6 +42,7 @@ import {
   ORIENTATION_TYPES,
   NODE_TYPES,
 } from "./tree.js";
+import { production } from "../shared/settings.js";
 
 /** @typedef {import('./extension.js').default} ForgeExtension */
 
@@ -61,6 +63,9 @@ export class WindowManager extends GObject.Object {
   constructor(ext) {
     super();
     this.ext = ext;
+    this.prefsTitle = `Forge ${_("Settings")} - ${
+      !production ? "DEV" : `${PACKAGE_VERSION}-${this.metadata.version}`
+    }`;
     this.windowProps = this.ext.configMgr.windowProps;
     this._kbd = this.ext.keybindings;
     this._tree = new Tree(this);
@@ -653,7 +658,7 @@ export class WindowManager extends GObject.Object {
         }
         break;
       case "PrefsOpen":
-        let existWindow = Utils.findWindowWith(this.ext.prefsTitle);
+        let existWindow = Utils.findWindowWith(this.prefsTitle);
         if (existWindow && existWindow.get_workspace()) {
           existWindow
             .get_workspace()
@@ -1484,7 +1489,7 @@ export class WindowManager extends GObject.Object {
   postProcessWindow(nodeWindow) {
     let metaWindow = nodeWindow.nodeValue;
     if (metaWindow) {
-      if (metaWindow.get_title() === this.ext.prefsTitle) {
+      if (metaWindow.get_title() === this.prefsTitle) {
         metaWindow
           .get_workspace()
           .activate_with_focus(metaWindow, global.display.get_current_time());
