+++ lib/prefs/keyboard.js
@@ -8,6 +8,7 @@ import { gettext as _ } from "resource:///org/gnome/Shell/Extensions/js/extensio
 
 // Prefs UI
 import { EntryRow, PreferencesPage, RadioRow } from "./widgets.js";
+import { Logger } from "../shared/logger.js";
 
 export class KeyboardPage extends PreferencesPage {
   static {
@@ -77,8 +78,6 @@ export class KeyboardPage extends PreferencesPage {
                   return Gtk.accelerator_valid(key, mods) && Gtk.accelerator_name(key, mods);
                 });
                 if (mappings.every((x) => !!x)) {
-                  // TODO: Logger is on the instance so far, so will need a
-                  // refactor
                   Logger.info("setting", bind, "to", mappings);
                   settings.set_strv(bind, mappings);
                 }
